+++ my2/Backend/src/main/java/service/ZamestnanecEntityFacadeREST.java
@@ -37,20 +37,20 @@ public class ZamestnanecEntityFacadeREST extends AbstractFacade<ZamestnanecEntit
 	@PUT
 	@Path("{id}")
 	@Consumes({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-	public void edit(@PathParam("id") Long id, ZamestnanecEntity entity) {
+	public void edit(@PathParam("id") Integer id, ZamestnanecEntity entity) {
 		super.edit(entity);
 	}
 
 	@DELETE
 	@Path("{id}")
-	public void remove(@PathParam("id") Long id) {
+	public void remove(@PathParam("id") Integer id) {
 		super.remove(super.find(id));
 	}
 
 	@GET
 	@Path("{id}")
 	@Produces({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-	public ZamestnanecDTO findZamestnanec(@PathParam("id") Long id) {
+	public ZamestnanecDTO findZamestnanec(@PathParam("id") Integer id) {
 		ZamestnanecDTO zamestnanec = new ZamestnanecDTO();
 		ArrayList<ZamestnanecEntity> res = new ArrayList<>();
 		res.add(super.find(id));
